Adding a border for the best discard
The border shows the best tile one should have discarded.

This is shown in the history box for the most efficent tile,
but this also makes it visible in your hand.

Further improvements can show values of available tiles by selection
and shanten as borders.  So we can see the distance between efficency.